Struct aws_sdk_mediatailor::client::fluent_builders::PutPlaybackConfiguration

pub struct PutPlaybackConfiguration<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> PutPlaybackConfiguration<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<PutPlaybackConfigurationOutput, SdkError<PutPlaybackConfigurationError>> where
R::Policy: SmithyRetryPolicy<PutPlaybackConfigurationInputOperationOutputAlias, PutPlaybackConfigurationOutput, PutPlaybackConfigurationError, PutPlaybackConfigurationInputOperationRetryAlias>,

pub fn ad_decision_server_url(self, inp: impl Into<String>) -> Self

The URL for the ad decision server (ADS). This includes the specification of static parameters and placeholders for dynamic parameters. AWS Elemental MediaTailor substitutes player-specific and session-specific parameters as needed when calling the ADS. Alternately, for testing you can provide a static VAST URL. The maximum length is 25,000 characters.

pub fn set_ad_decision_server_url(self, input: Option<String>) -> Self

pub fn avail_suppression(self, inp: AvailSuppression) -> Self

The configuration for avail suppression, also known as ad suppression. For more information about ad suppression, see Ad Suppression.

pub fn set_avail_suppression(self, input: Option<AvailSuppression>) -> Self

pub fn bumper(self, inp: Bumper) -> Self

The configuration for bumpers. Bumpers are short audio or video clips that play at the start or before the end of an ad break. To learn more about bumpers, see Bumpers.

pub fn set_bumper(self, input: Option<Bumper>) -> Self

pub fn cdn_configuration(self, inp: CdnConfiguration) -> Self

The configuration for using a content delivery network (CDN), like Amazon CloudFront, for content and ad segment management.

pub fn set_cdn_configuration(self, input: Option<CdnConfiguration>) -> Self

pub fn configuration_aliases(
self,
k: impl Into<String>,
v: impl Into<HashMap<String, String>>
) -> Self

Adds a key-value pair to ConfigurationAliases.

To override the contents of this collection use set_configuration_aliases.

The player parameters and aliases used as dynamic variables during session initialization. For more information, see Domain Variables.

pub fn set_configuration_aliases(
self,
input: Option<HashMap<String, HashMap<String, String>>>
) -> Self

pub fn dash_configuration(self, inp: DashConfigurationForPut) -> Self

The configuration for DASH content.

pub fn set_dash_configuration(
self,
input: Option<DashConfigurationForPut>
) -> Self

pub fn live_pre_roll_configuration(self, inp: LivePreRollConfiguration) -> Self

The configuration for pre-roll ad insertion.

pub fn set_live_pre_roll_configuration(
self,
input: Option<LivePreRollConfiguration>
) -> Self

pub fn manifest_processing_rules(self, inp: ManifestProcessingRules) -> Self

The configuration for manifest processing rules. Manifest processing rules enable customization of the personalized manifests created by MediaTailor.

pub fn set_manifest_processing_rules(
self,
input: Option<ManifestProcessingRules>
) -> Self

pub fn name(self, inp: impl Into<String>) -> Self

The identifier for the playback configuration.

pub fn set_name(self, input: Option<String>) -> Self

pub fn personalization_threshold_seconds(self, inp: i32) -> Self

Defines the maximum duration of underfilled ad time (in seconds) allowed in an ad break. If the duration of underfilled ad time exceeds the personalization threshold, then the personalization of the ad break is abandoned and the underlying content is shown. This feature applies to ad replacement in live and VOD streams, rather than ad insertion, because it relies on an underlying content stream. For more information about ad break behavior, including ad replacement and insertion, see Ad Behavior in AWS Elemental MediaTailor.

pub fn set_personalization_threshold_seconds(self, input: Option<i32>) -> Self

pub fn slate_ad_url(self, inp: impl Into<String>) -> Self

The URL for a high-quality video asset to transcode and use to fill in time that's not used by ads. AWS Elemental MediaTailor shows the slate to fill in gaps in media content. Configuring the slate is optional for non-VPAID configurations. For VPAID, the slate is required because MediaTailor provides it in the slots that are designated for dynamic ad content. The slate must be a high-quality asset that contains both audio and video.

pub fn set_slate_ad_url(self, input: Option<String>) -> Self

pub fn tags(self, k: impl Into<String>, v: impl Into<String>) -> Self

Adds a key-value pair to Tags.

To override the contents of this collection use set_tags.

The tags to assign to the playback configuration.

pub fn set_tags(self, input: Option<HashMap<String, String>>) -> Self

pub fn transcode_profile_name(self, inp: impl Into<String>) -> Self

The name that is used to associate this playback configuration with a custom transcode profile. This overrides the dynamic transcoding defaults of MediaTailor. Use this only if you have already set up custom profiles with the help of AWS Support.

pub fn set_transcode_profile_name(self, input: Option<String>) -> Self

pub fn video_content_source_url(self, inp: impl Into<String>) -> Self

The URL prefix for the parent manifest for the stream, minus the asset ID. The maximum length is 512 characters.

pub fn set_video_content_source_url(self, input: Option<String>) -> Self