Module aws_sdk_mediatailor::model

Modules

access_configuration

See AccessConfiguration

ad_break

See AdBreak

ad_marker_passthrough

See AdMarkerPassthrough

alert

See Alert

avail_suppression

See AvailSuppression

bumper

See Bumper

cdn_configuration

See CdnConfiguration

channel

See Channel

dash_configuration

See DashConfiguration

dash_configuration_for_put

See DashConfigurationForPut

dash_playlist_settings

See DashPlaylistSettings

default_segment_delivery_configuration

See DefaultSegmentDeliveryConfiguration

hls_configuration

See HlsConfiguration

hls_playlist_settings

See HlsPlaylistSettings

http_configuration

See HttpConfiguration

http_package_configuration

See HttpPackageConfiguration

live_pre_roll_configuration

See LivePreRollConfiguration

manifest_processing_rules

See ManifestProcessingRules

playback_configuration

See PlaybackConfiguration

request_output_item

See RequestOutputItem

response_output_item

See ResponseOutputItem

schedule_ad_break

See ScheduleAdBreak

schedule_configuration

See ScheduleConfiguration

schedule_entry

See ScheduleEntry

secrets_manager_access_token_configuration

See SecretsManagerAccessTokenConfiguration

slate_source

See SlateSource

source_location

See SourceLocation

splice_insert_message

See SpliceInsertMessage

transition

See Transition

vod_source

See VodSource

Structs

AccessConfiguration

Access configuration parameters.

AdBreak

Ad break configuration parameters.

AdMarkerPassthrough

For HLS, when set to true, MediaTailor passes through EXT-X-CUE-IN, EXT-X-CUE-OUT, and EXT-X-SPLICEPOINT-SCTE35 ad markers from the origin manifest to the MediaTailor personalized manifest.

No logic is applied to these ad markers. For example, if EXT-X-CUE-OUT has a value of 60, but no ads are filled for that ad break, MediaTailor will not set the value to 0.

Alert

Alert configuration parameters.

AvailSuppression

The configuration for avail suppression, also known as ad suppression. For more information about ad suppression, see Ad Suppression.

Bumper

The configuration for bumpers. Bumpers are short audio or video clips that play at the start or before the end of an ad break. To learn more about bumpers, see Bumpers.

CdnConfiguration

The configuration for using a content delivery network (CDN), like Amazon CloudFront, for content and ad segment management.

Channel

The configuration parameters for a channel.

DashConfiguration

The configuration for DASH content.

DashConfigurationForPut

The configuration for DASH PUT operations.

DashPlaylistSettings

Dash manifest configuration parameters.

DefaultSegmentDeliveryConfiguration

The optional configuration for a server that serves segments. Use this if you want the segment delivery server to be different from the source location server. For example, you can configure your source location server to be an origination server, such as MediaPackage, and the segment delivery server to be a content delivery network (CDN), such as CloudFront. If you don't specify a segment delivery server, then the source location server is used.

HlsConfiguration

The configuration for HLS content.

HlsPlaylistSettings

HLS playlist configuration parameters.

HttpConfiguration

The HTTP configuration for the source location.

HttpPackageConfiguration

The HTTP package configuration properties for the requested VOD source.

LivePreRollConfiguration

The configuration for pre-roll ad insertion.

ManifestProcessingRules

The configuration for manifest processing rules. Manifest processing rules enable customization of the personalized manifests created by MediaTailor.

PlaybackConfiguration

Creates a playback configuration. For information about MediaTailor configurations, see Working with configurations in AWS Elemental MediaTailor.

RequestOutputItem

The output configuration for this channel.

ResponseOutputItem

This response includes only the "property" : "type" property.

ScheduleAdBreak

The schedule's ad break properties.

ScheduleConfiguration

Schedule configuration parameters. A channel must be stopped before changes can be made to the schedule.

ScheduleEntry

The properties for a schedule.

SecretsManagerAccessTokenConfiguration

AWS Secrets Manager access token configuration parameters. For information about Secrets Manager access token authentication, see Working with AWS Secrets Manager access token authentication.

SlateSource

Slate VOD source configuration.

SourceLocation

This response includes only the "type" : "object" property.

SpliceInsertMessage

Splice insert message configuration.

Transition

Program transition configuration.

VodSource

VOD source configuration parameters.

Enums

AccessType

ChannelState

MessageType

Mode

OriginManifestType

PlaybackMode

RelativePosition

ScheduleEntryType

Type