Module aws_sdk_mediatailor::operation

Structs

CreateChannel

Creates a channel.

CreateProgram

Creates a program.

CreateSourceLocation

Creates a source location on a specific channel.

CreateVodSource

Creates name for a specific VOD source in a source location.

DeleteChannel

Deletes a channel. You must stop the channel before it can be deleted.

DeleteChannelPolicy

Deletes a channel's IAM policy.

DeletePlaybackConfiguration

Deletes the playback configuration for the specified name.

DeleteProgram

Deletes a specific program on a specific channel.

DeleteSourceLocation

Deletes a source location on a specific channel.

DeleteVodSource

Deletes a specific VOD source in a specific source location.

DescribeChannel

Describes the properties of a specific channel.

DescribeProgram

Retrieves the properties of the requested program.

DescribeSourceLocation

Retrieves the properties of the requested source location.

DescribeVodSource

Provides details about a specific VOD source in a specific source location.

GetChannelPolicy

Retrieves information about a channel's IAM policy.

GetChannelSchedule

Retrieves information about your channel's schedule.

GetPlaybackConfiguration

Returns the playback configuration for the specified name.

ListAlerts

Returns a list of alerts for the given resource.

ListChannels

Retrieves a list of channels that are associated with this account.

ListPlaybackConfigurations

Returns a list of the playback configurations defined in AWS Elemental MediaTailor. You can specify a maximum number of configurations to return at a time. The default maximum is 50. Results are returned in pagefuls. If MediaTailor has more configurations than the specified maximum, it provides parameters in the response that you can use to retrieve the next pageful.

ListSourceLocations

Retrieves a list of source locations.

ListTagsForResource

Returns a list of the tags assigned to the specified playback configuration resource.

ListVodSources

Lists all the VOD sources in a source location.

PutChannelPolicy

Creates an IAM policy for the channel.

PutPlaybackConfiguration

Adds a new playback configuration to AWS Elemental MediaTailor.

StartChannel

Starts a specific channel.

StopChannel

Stops a specific channel.

TagResource

Adds tags to the specified playback configuration resource. You can specify one or more tags to add.

UntagResource

Removes tags from the specified playback configuration resource. You can specify one or more tags to remove.

UpdateChannel

Updates an existing channel.

UpdateSourceLocation

Updates a source location on a specific channel.

UpdateVodSource

Updates a specific VOD source in a specific source location.